build(all): pretty much overhauled the whole build system
- Added debug mode support.
- Scripts now take arguments.
- Run script now correctly passes arguments to the runtime executable.
- Run and build script now reconfigures the CMake project.
- CMake now correctly handles common as a library (I hope).

BREAKING CHANGE: The way every CMake project handles the common dependencies changed fundamentally.
I think I fixed it, but just in case, include absolutely relatively to the include directory (e.g,
prefer 'common/geometry/cone.hpp' instead of 'geometry/cone.hpp' when in the same namespace).

+++ clean.sh
@@ -1,1 +1,24 @@
-rm -r build
+#!/bin/bash
+
+###################################################################################
+# clean.sh
+# Cleans the target directory.
+#
+# FLAGS:
+#   --debug, -d: Builds the project in debug mode.
+###################################################################################
+
+source .build/args.sh
+
+OUTDIR="build/out"
+
+if has_param 1 '--all' "$@"; then
+    echo "NUKING THIS SHIT"
+    if [ -d "build" ] ; then
+        rm -r build
+    fi
+else
+    if [ -d "$OUTDIR" ] ; then
+        rm -r $OUTDIR
+    fi
+fi
